feat: move CookieConsentConfig.tsx to utils
